feature: Auto refresh data when connection changes, move auto refreshing into viewmodel and cancel in DisposableEffect

+++ app/src/main/java/eu/wedgess/mihole/data/PiHoleRepository.kt
@@ -156,7 +156,7 @@ class PiHoleRepository @Inject constructor(
         }
     }
 
-    suspend fun updateRefreshInterval(interval: Int) = kotlin.runCatching {
+    suspend fun updateRefreshInterval(interval: Long) = kotlin.runCatching {
         userPreferences.updateData { preferences ->
             preferences.toBuilder().setRefreshTime(interval).build()
         }
